+++ SQLiteStudio3/guiSQLiteStudio/dataview.cpp
@@ -589,6 +589,7 @@ void DataView::togglePerColumnFiltering()
     perColumnAreaParent->setVisible(enable);
 
     recreateFilterInputs();
+    applyFilter();
 }
 
 void DataView::updateCommitRollbackActions(bool enabled)
@@ -819,11 +820,15 @@ void DataView::applyFilter()
 
     if (actionMap[FILTER_PER_COLUMN]->isChecked())
     {
-        filterValues.clear();
+        QStringList newValues;
         for (QLineEdit* edit : filterInputs)
-            filterValues << edit->text();
+            newValues << edit->text();
 
-        if (filterValues.join("").isEmpty())
+        if (newValues == lastColumnFilterValues)
+            return;
+
+        lastColumnFilterValues = newValues;
+        if (lastColumnFilterValues.join("").isEmpty())
         {
             model->resetFilter();
             return;
@@ -831,24 +836,26 @@ void DataView::applyFilter()
 
         switch (filterMode)
         {
-            case DataView::FilterMode::STRING:
-                model->applyStringFilter(filterValues);
-                break;
             case DataView::FilterMode::SQL:
-                // Should never happen.
-                qWarning() << "Requested to filter by SQL for filtering per-column. This should not be possible.";
+                // If use had SQL selected prior to enabling per-column filtering, we filter just like it was a STRING.
+            case DataView::FilterMode::STRING:
+                model->applyStringFilter(lastColumnFilterValues);
                 break;
             case DataView::FilterMode::REGEXP:
-                model->applyRegExpFilter(filterValues);
+                model->applyRegExpFilter(lastColumnFilterValues);
                 break;
             case FilterMode::EXACT:
-                model->applyStrictFilter(filterValues);
+                model->applyStrictFilter(lastColumnFilterValues);
                 break;
         }
     }
     else
     {
         QString value = filterEdit->text();
+        if (value == lastSingleFilterValue)
+            return;
+
+        lastSingleFilterValue = value;
         switch (filterMode)
         {
             case DataView::FilterMode::STRING:
@@ -964,10 +971,10 @@ void DataView::recreateFilterInputs()
         edit->setClearButtonEnabled(true);
         edit->setFixedWidth(gridView->columnWidth(i));
         edit->setToolTip(tr("Hit Enter key or press \"Apply filter\" button on toolbar to apply new value."));
-        if (filterValues.size() > i)
-            edit->setText(filterValues[i]);
+        if (lastColumnFilterValues.size() > i)
+            edit->setText(lastColumnFilterValues[i]);
 
-        connect(edit, SIGNAL(returnPressed()), this, SLOT(applyFilter()));
+        connect(edit, SIGNAL(editingFinished()), this, SLOT(applyFilter()));
         perColumnWidget->layout()->addWidget(edit);
         filterInputs << edit;
     }